First commit
1. In feature_engineering.py,
Function "apply_min_max_scaler" is added for using MinMaxScaler to normalize feature which works best for LSTM after numerous testing. And it will save the file.

2. In modelling.py,
Two function for metrics (accuracy percentage, mse) are added.
Function "reconstruct"  is added for FNN. It convert all the label 2  to 0
Function "fit_FNN" is added for providing FNN model
Function "univariate_data" and "transform_and_rescale" are added for the scaling of label and the conversion to the sequential data.
Function "fit_LSTM" is added for providing LSTM model

In modelling.py,
Two blocks of codes with similar styles as before are added. The only difference is 1~2 lines of code are added for the "unique" data input for models.

In results.csv,
The result of FNN (classification) and LSTM (regression) are added.